services/vision/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
opencv-python-headless==4.10.0.84
python-chess==1.999
numpy==1.26.4
